automatecopywriter/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML dependencies
smolagents==0.1.0
litellm==1.17.0
openai==1.3.7

# Web scraping and HTTP requests
requests==2.31.0
httpx==0.28.1
ddgs==9.6.1

# Environment and configuration
python-dotenv==1.0.0

# Data processing and utilities
python-multipart==0.0.6
jinja2==3.1.2


# Security and validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
structlog==23.2.0

# Optional: Database support (uncomment if needed)
# sqlalchemy==2.0.23
# alembic==1.13.1
# psycopg2-binary==2.9.9

# Optional: Redis for caching (uncomment if needed)
# redis==5.0.1
# aioredis==2.0.1
